backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
paypalrestsdk==1.13.3
twilio==8.10.0
sib-api-v3-sdk==7.6.0
redis==5.0.1
websockets==12.0
aiofiles==23.2.1
pillow==10.1.0